cleaned up navbar and footer styling, finished user profile styling, removed unused css files
