+++ src/App.jsx
@@ -3,9 +3,9 @@ import React, { useState, useEffect } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import PageWrapper from './components/PageWrapper';
 
-import './App.css';
-
 import { getAllData } from './util/index';
+import LoginForm from "./components/LoginForm.jsx";
+import RegisterForm from "./components/RegisterForm.jsx";
 import Navbar from './components/landing/Navbar.jsx';
 import HeroSection from './components/landing/HeroSection.jsx';
 import HowSection from './components/landing/HowSection.jsx';
@@ -13,8 +13,7 @@ import TestimonialSection from './components/landing/TestimonialSection.jsx';
 import SignUpSection from  './components/landing/SignUpSection.jsx';
 import FooterSection from  './components/landing/FooterSection.jsx';
 import JobSearch from './components/JobSearch.jsx';
-import './App.css';
-import UserInfoForm from "./components/UserInfoForm.jsx";
+import UserInfoForm from "./components/UserProfile/UserInfoForm.jsx";
