--- src/components/UserProfile/userProfile.module.css
@@ -1,25 +1,0 @@
-.profileContainer {
-  padding: 2rem;
-  max-width: 600px;
-  margin: 0 auto;
-}
-
-.profileCard {
-  background-color: #fff;
-  padding: 2rem;
-  border-radius: 12px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-}
-
-.profileCard p {
-  margin: 0.5rem 0;
-  font-size: 1rem;
-}
-
-.profileImage {
-  width: 120px;
-  height: 120px;
-  border-radius: 50%;
-  object-fit: cover;
-  margin-bottom: 1rem;
-}
