+++ src/routeStyles.js
@@ -1,0 +1,18 @@
+export const navbarRouteStyles = {
+    "/": { navbar: "bg-ny-pink", button: "btn-grn", text: "text-white" },
+    "/login": { navbar: "bg-monte-carlo", button: "btn-pnk", text: "text-black" },
+    "/register": { navbar: "bg-monte-carlo", button: "btn-pnk", text: "text-black" },
+    "/profile": { navbar: "bg-monte-carlo", button: "btn-blk", text: "text-black" },
+    "/edit-profile": { navbar: "bg-monte-carlo", button: "btn-blk", text: "text-black" },
+    "/jobs": { navbar: "bg-monte-carlo-dark", button: "btn-blk", text: "text-black" },
+    "/post-job": { navbar: "bg-white", button: "btn-blk", text: "text-black" },
+};
+
+export const footerRouteStyles = {
+    "/": { bg: "bg-ny-pink-light", text: "text-black", line: "border-gray-400" },
+    "/login": { bg: "bg-ny-pink-light", text: "text-black", line: "border-gray-400" },
+    "/register": { bg: "bg-ny-pink-light", text: "text-black", line: "border-gray-400" },
+    "/profile": { bg: "bg-ny-pink", text: "text-white", line: "line-white" },
+    "/edit-profile": { bg: "bg-ny-pink", text: "text-white", line: "line-white" },
+    "/jobs": { bg: "bg-ny-pink", text: "text-white", line: "line-white" },
+};
